database: Use `auth_user` directive as intended
Prior to this commit, the `auth_user` directive was translated as `user`
directive. Those directives are not the same.

The `user` directive expects credentials to be solved either by the userlist.txt
mechanism or by a password provided in the database entry.

The `auth_user` directive allows the credentials to be solved through querying.

Ultimately, it must be the module's user to decide which to use.

Related #31